emailerlogix/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
environ==1.0
idna==3.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.1
pika==1.3.1
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
